+++ shuup/admin/modules/orders/__init__.py
@@ -14,6 +14,7 @@ from django.utils.translation import ugettext_lazy as _
 
 from shuup.admin.base import AdminModule, MenuEntry, Notification, SearchResult
 from shuup.admin.menu import ORDERS_MENU_CATEGORY, STOREFRONT_MENU_CATEGORY
+from shuup.admin.shop_provider import get_shop
 from shuup.admin.utils.permissions import (
     get_default_model_permissions, get_permissions_from_urls
 )
@@ -180,7 +181,9 @@ class OrderModule(AdminModule):
                 )
 
     def get_notifications(self, request):
+        shop = get_shop(request)
         old_open_orders = Order.objects.filter(
+            shop=shop,
             status__role=OrderStatusRole.INITIAL,
             order_date__lt=now() - timedelta(days=4)
         ).count()