+++ shuup/admin/modules/services/forms.py
@@ -14,6 +14,7 @@ from django_countries import Countries
 
 from shuup.admin.forms import ShuupAdminForm
 from shuup.admin.forms.widgets import TextEditorWidget
+from shuup.admin.shop_provider import get_shop
 from shuup.core.models import (
     CountryLimitBehaviorComponent, FixedCostBehaviorComponent,
     GroupAvailabilityBehaviorComponent, OrderTotalLimitBehaviorComponent,
@@ -27,10 +28,10 @@ class BaseMethodForm(ShuupAdminForm):
         model = None
         exclude = [
             "identifier", "behavior_components", "old_module_identifier",
-            "old_module_data"
+            "old_module_data", "shop"
         ]
         base_fields = [
-            "choice_identifier", "name", "description", "enabled", "shop",
+            "choice_identifier", "name", "description", "enabled",
             "logo", "tax_class"
         ]
         widgets = {
@@ -66,6 +67,7 @@ class BaseMethodForm(ShuupAdminForm):
         setattr(self.instance, self.service_provider_attr, value)
 
     def _save_master(self, commit=True):
+        self.cleaned_data['shop'] = get_shop(self.request)
         if self.instance.pk:
             return super(BaseMethodForm, self)._save_master(commit)