+++ shuup/admin/views/select.py
@@ -17,6 +17,7 @@ from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
 from django.views.generic import TemplateView
 
+from shuup.admin.shop_provider import get_shop
 from shuup.core.models import Carrier, Contact, Product
 
 
@@ -75,7 +76,10 @@ class MultiselectAjaxView(TemplateView):
         cls = apps.get_model(model_name)
         qs = cls.objects.all()
         if hasattr(cls.objects, "all_except_deleted"):
-            qs = cls.objects.all_except_deleted()
+            shop = get_shop(self.request) if not self.request.user.is_superuser else None
+            qs = cls.objects.all_except_deleted(shop=shop)
+        if hasattr(cls.objects, "get_for_user"):
+            qs = cls.objects.get_for_user(self.request.user)
         self.init_search_fields(cls)
         if not self.search_fields:
             return [{"id": None, "name": _("Couldn't get selections for %s.") % model_name}]